requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymupdf==1.24.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
google-generativeai==0.7.2
python-multipart==0.0.9
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.1